pyproject.toml
[project]
name = "django-tasks-cloud"
version = "0.1.0b2"
authors = [
  { name="Developers at PaperCloud Tech", email="developers@papercloud.tech" },
]
description = "Manage long running tasks out of request-response cycle with Django's all-new `Task` back-end and AWS/Azure native-services."
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

dependencies = [
    "django==6.0",
]

[dependency-groups]
aws = [
    "boto3>=1.42.8",
]
azure = [
    "azure-identity>=1.25.1",
    "azure-servicebus>=7.14.3",
    "azure-storage-queue>=12.14.1",
]
dev = [
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["uv_build >= 0.9.17, <0.10.0"]
build-backend = "uv_build"

[project.urls]
Homepage = "https://github.com/papercloudtech/django-cloud-tasks"
Issues = "https://github.com/papercloudtech/django-cloud-tasks/issues"
